requirements.txt
# YOLOv8 and training
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Feature extraction and matching
git+https://github.com/openai/CLIP.git
transformers>=4.30.0
pillow>=9.0.0

# Vector search
faiss-cpu
# faiss-gpu  # Uncomment if you have CUDA GPU

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6